Remove `WithExceptionHandler` traits
