--- engine/flink/components-api/src/main/scala/pl/touk/nussknacker/engine/flink/api/exception/SinkWriterWithExceptionHandler.scala
@@ -1,16 +1,0 @@
-package pl.touk.nussknacker.engine.flink.api.exception
-
-import org.apache.flink.api.connector.sink2.SinkWriter
-
-trait SinkWriterWithExceptionHandler[T] {
-  self: SinkWriter[T] =>
-
-  protected val exceptionHandler: ExceptionHandler
-
-  override def close(): Unit = {
-    if (exceptionHandler != null) {
-      exceptionHandler.close()
-    }
-  }
-
-}
